+++ src/main/java/com/techelevator/items/DrinkItem.java
@@ -1,11 +1,13 @@
 package com.techelevator.items;
 
+import java.math.BigDecimal;
+
 public class DrinkItem extends Item {
 
     public static final String PRODUCT_TYPE_NAME = "Drink";
     public static final String DISPENSE_MESSAGE = "Glug Glug, Yum!";
 
-    public DrinkItem(String slotLocation, String productName, double price, int quantityInStock) {
+    public DrinkItem(String slotLocation, String productName, BigDecimal price, int quantityInStock) {
         super(slotLocation, productName, price, quantityInStock);
     }
