+++ src/main/java/com/techelevator/items/GumItem.java
@@ -1,11 +1,13 @@
 package com.techelevator.items;
 
+import java.math.BigDecimal;
+
 public class GumItem extends Item {
 
     public static final String PRODUCT_TYPE_NAME = "Gum";
     public static final String DISPENSE_MESSAGE = "Chew Chew, Yum!";
 
-    public GumItem(String slotLocation, String productName, double price, int quantityInStock) {
+    public GumItem(String slotLocation, String productName, BigDecimal price, int quantityInStock) {
         super(slotLocation, productName, price, quantityInStock);
     }
