+++ src/main/java/com/techelevator/util/ActionLogger.java
@@ -3,6 +3,7 @@ package com.techelevator.util;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.PrintWriter;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
@@ -12,7 +13,7 @@ public class ActionLogger {
     private static final String LOG_ENTRY_FORMAT = "%s %s %s: $%.2f $%.2f" + System.lineSeparator();
     private static PrintWriter logFileWriter;
 
-    public static void log(String action, double amount, double updatedCurrentMoney) {
+    public static void log(String action, BigDecimal amount, BigDecimal updatedCurrentMoney) {
         try {
             if (logFileWriter == null) {
                 File logFile = new File(LOG_FILE_PATH);
@@ -23,8 +24,8 @@ public class ActionLogger {
                     now.format(DateTimeFormatter.ofPattern("MM/dd/yyyy")),
                     now.format(DateTimeFormatter.ofLocalizedTime(FormatStyle.MEDIUM)),
                     action,
-                    amount,
-                    updatedCurrentMoney);
+                    amount.doubleValue(),
+                    updatedCurrentMoney.doubleValue());
             logFileWriter.flush();
         }
         catch (Exception e) {
